Access information in a cvo object

Utility functions for cvo object. More in section "Details."

cvo_get_inds(cvo, fold, type = c("asis", "train", "test"))

# S3 method for cvo_caret
cvo_get_inds(cvo, fold, type = c("asis", "train", "test"))

# S3 method for cvo_mlr
cvo_get_inds(cvo, fold, type = c("asis", "train", "test"))

cvo_get_info(cvo)

cvo_get_sample_size(cvo)

cvo_get_seeds(cvo)

cvo_count_folds(cvo)

# S3 method for cvo_caret
cvo_count_folds(cvo)

# S3 method for cvo_mlr
cvo_count_folds(cvo)

cvo_get_fold_names(cvo)

# S3 method for cvo_caret
cvo_get_fold_names(cvo)

# S3 method for cvo_mlr
cvo_get_fold_names(cvo)

Arguments

cvo
a cross-validation object (cvo) created with function cvo_create_folds.

fold
either a name or an index of fold of interest.

type
(string) one of options, indicating the kind of indices you are interested in:

"AsIs" - indices as they are in the object,

"Train" - indices of training set,

"Test" - indices of test set.

Details

Function cvo_get_info returns information about the cross-validation object cvo.

Function cvo_get_sample_size returns sample size of the object used to create the cross-validation object cvo.

Function cvo_count_folds returns total number of folds in a cvo object.

Function cvo_get_inds extracts indices of indicated set (either training or test) from cross-validation object (cvo) created with function cvo_create_folds when fold of interest is indicated as fold (it can be either a name of fold or an index of fold). If cvo does not contain indices of chosen type, it returns a complement to those indices and this information is indicated in attributes of returned object.

Function cvo_get_seeds returns information about seeds of (pseudo)random number generator used for each repetition of splitting to folds.

See also

cvo_create_folds

Examples

cvo <- cvo_create_folds(fluorescence, folds = 10)

#> Error in set.seed(seeds[i], kind = kind): supplied seed is not a valid integer

cvo

#> $Rep1_Fold1
#>  [1]  1  2  3  4  5  6  7  8 12 13 14 15 16 17 18 20 21 22 23 24 26 27 28 29
#> 
#> $Rep1_Fold2
#>  [1]  1  4  5  6  7  8  9 10 11 12 13 14 15 17 18 19 23 24 25 26 27 28 29 30
#> 
#> $Rep1_Fold3
#>  [1]  1  2  3  4  7  8  9 10 11 13 15 16 17 18 19 20 21 22 23 24 25 26 27 30
#> 
#> $Rep1_Fold4
#>  [1]  2  3  4  5  6  8  9 10 11 12 14 16 17 18 19 20 21 22 23 25 27 28 29 30
#> 
#> $Rep1_Fold5
#>  [1]  1  2  3  5  6  7  9 10 11 12 13 14 15 16 19 20 21 22 24 25 26 28 29 30
#> 
#> attr(,"class")
#> [1] "cvo_caret" "cvo"      
#> attr(,"info")
#>   indices stratified blocked cv_type k repetitions sample_size
#> 1   Train       TRUE   FALSE  k-fold 5           1          30
#> attr(,"seeds")
#> attr(,"seeds")$generator
#> [1] "L'Ecuyer-CMRG"
#> 
#> attr(,"seeds")$seeds
#> [1] 39100040
#> 

cvo_get_info(cvo)

#>   indices stratified blocked cv_type k repetitions sample_size
#> 1   Train       TRUE   FALSE  k-fold 5           1          30

cvo_get_seeds(cvo)

#> $generator
#> [1] "L'Ecuyer-CMRG"
#> 
#> $seeds
#> [1] 39100040
#> 

cvo_get_sample_size(cvo)

#> [1] 30

cvo_count_folds(cvo)

#> [1] 5

cvo_get_fold_names(cvo)

#> [1] "Rep1_Fold1" "Rep1_Fold2" "Rep1_Fold3" "Rep1_Fold4" "Rep1_Fold5"

cvo_get_inds(cvo, 1)

#>  [1]  1  2  3  4  5  6  7  8 12 13 14 15 16 17 18 20 21 22 23 24 26 27 28 29
#> attr(,"fold")
#> [1] "Rep1_Fold1"
#> attr(,"indices")
#> [1] "Train"

cvo_get_inds(cvo, 1, "train")

#>  [1]  1  2  3  4  5  6  7  8 12 13 14 15 16 17 18 20 21 22 23 24 26 27 28 29
#> attr(,"fold")
#> [1] "Rep1_Fold1"
#> attr(,"indices")
#> [1] "train"

cvo_get_inds(cvo, 1, "test")

#> [1]  9 10 11 19 25 30
#> attr(,"fold")
#> [1] "Rep1_Fold1"
#> attr(,"indices")
#> [1] "Complement to Train"