[!] Carry the ROC analysis

DO the ROC (receiver operating characteristic) analysis and calculate vector of cut-off values and associated number of true positives (TP), false negatives (FN), false positives (FP), and true negatives (TN) as well as performance measures such as sensitivity, specificity, etc. [!!!]

print.roc_results(x, ...)

print.roc_optimal(x, digits = 3, ...)

print.roc_info(x, ...)

roc_analysis(x, gr, pos_label = levels(gr)[2], pos_is_larger = NULL,
  optimize_by = "bac", results = "all")

Arguments

x
(in print function) An object to print.

...
[!!!] Passed to further methods.

gr
(factor) A factor vector with two levels.

pos_label
(character(1)) A string with the name of positive group.

pos_is_larger
(NULL|TRUE|FALSE) A flag indicating, if values of positive group are (on average) larger. If NULL, this option is determined basing on data using 10% trimmed mean.

optimize_by
(string(1)) [!!!] Method to determine the optimal cut-off value. Current options: "bac", "youden", "kappa".

results
(character(1)) A string indicating which results should be returned: either "all" (as described in section "Values") or just "optimal".

x
(numeric) A numeric vector.

Value

A list (which also inherits from class "roc_result_list") with three fields: $info, $optimal, $all_results. $info is a data frame with columns var_name - empty string reserved for variable name, neg_label, pos_label labels of negative and positive groups respectively, n_neg, n_pos, n_total - number of negative and positive cases as well as number of cases in total. $optimal one row from $all_results, which was determined as having optima threshold (cut-off) value. Sometimes it can be several rows, if the performance is equally good. $all_results is a data frame with columns cutoffs for cutoff values, tp (number of true positives), fn (number of false negatives), fp (number of false positives), tn (number of true negatives), ... [!!!]

Note

This function is based on functions predict and .compute.unnormalized.roc.curve from ROCR package.

See also

Other functions for ROC: access_elements, roc_performance_measures

Examples

# Make some data
set.seed(1)
(x_ <- rnorm(10))

#>  [1] -0.6264538  0.1836433 -0.8356286  1.5952808  0.3295078 -0.8204684
#>  [7]  0.4874291  0.7383247  0.5757814 -0.3053884

(gr_ <- gl(n = 2, k = 5, length = 10, labels = c("H","S")))

#>  [1] H H H H H S S S S S
#> Levels: H S

# Explore the functions

roc_analysis(x_, gr_)

#> $info
#>  var_name neg_label n_neg pos_label n_pos n_total
#>                   H     5         S     5      10
#> 
#> $optimal
#>  cutoff tp fn fp tn sens spec  ppv   npv bac youden kappa auc
#>   0.487  3  2  1  4  0.6  0.8 0.75 0.667 0.7    0.4   0.4 0.6
#> 
#> * The optimal cut-off value selected by: max BAC
#> 
#> $all_results
#>     cutoff  tp  fn  fp  tn sens spec   ppv   npv bac youden
#> 1      Inf   0   5   0   5    0    1   NaN   0.5 0.5      0
#> 2      1.6   0   5   1   4    0  0.8     0 0.444 0.4   -0.2
#> 3    0.738   1   4   1   4  0.2  0.8   0.5   0.5 0.5      0
#> 4    0.576   2   3   1   4  0.4  0.8 0.667 0.571 0.6    0.2
#> ...    ... ... ... ... ...  ...  ...   ...   ... ...    ...
#> 8   -0.305   4   1   3   2  0.8  0.4 0.571 0.667 0.6    0.2
#> 9   -0.626   4   1   4   1  0.8  0.2   0.5   0.5 0.5      0
#> 10   -0.82   5   0   4   1    1  0.2 0.556     1 0.6    0.2
#> 11  -0.836   5   0   5   0    1    0   0.5   NaN 0.5      0
#> 
#> attr(,"class")
#> [1] "roc_result_list" "list"           

roc_analysis(x_, gr_, pos_label = "H")

#> $info
#>  var_name neg_label n_neg pos_label n_pos n_total
#>                   S     5         H     5      10
#> 
#> $optimal
#>  cutoff tp fn fp tn sens spec   ppv  npv bac youden kappa auc
#>    0.33  4  1  2  3  0.8  0.6 0.667 0.75 0.7    0.4   0.4 0.6
#> 
#> * The optimal cut-off value selected by: max BAC
#> 
#> $all_results
#>     cutoff  tp  fn  fp  tn sens spec   ppv   npv bac youden
#> 1     -Inf   0   5   0   5    0    1   NaN   0.5 0.5      0
#> 2   -0.836   1   4   0   5  0.2    1     1 0.556 0.6    0.2
#> 3    -0.82   1   4   1   4  0.2  0.8   0.5   0.5 0.5      0
#> 4   -0.626   2   3   1   4  0.4  0.8 0.667 0.571 0.6    0.2
#> ...    ... ... ... ... ...  ...  ...   ...   ... ...    ...
#> 8    0.487   4   1   3   2  0.8  0.4 0.571 0.667 0.6    0.2
#> 9    0.576   4   1   4   1  0.8  0.2   0.5   0.5 0.5      0
#> 10   0.738   4   1   5   0  0.8    0 0.444     0 0.4   -0.2
#> 11     1.6   5   0   5   0    1    0   0.5   NaN 0.5      0
#> 
#> attr(,"class")
#> [1] "roc_result_list" "list"