Access information in a cvo object

Utility functions for cvo object. More in section "Details."

cvo_get_inds(cvo, fold, type = c("asis", "train", "test"))

# S3 method for cvo_caret
cvo_get_inds(cvo, fold, type = c("asis", "train", "test"))

# S3 method for cvo_mlr
cvo_get_inds(cvo, fold, type = c("asis", "train", "test"))

cvo_get_info(cvo)

cvo_get_sample_size(cvo)

cvo_get_seeds(cvo)

cvo_count_folds(cvo)

# S3 method for cvo_caret
cvo_count_folds(cvo)

# S3 method for cvo_mlr
cvo_count_folds(cvo)

cvo_get_fold_names(cvo)

# S3 method for cvo_caret
cvo_get_fold_names(cvo)

# S3 method for cvo_mlr
cvo_get_fold_names(cvo)

Arguments

cvo
a cross-validation object (cvo) created with function cvo_create_folds.

fold
either a name or an index of fold of interest.

type
(string) one of options, indicating the kind of indices you are interested in:

"AsIs" - indices as they are in the object,

"Train" - indices of training set,

"Test" - indices of test set.

Details

Function cvo_get_info returns information about the cross-validation object cvo.

Function cvo_get_sample_size returns sample size of the object used to create the cross-validation object cvo.

Function cvo_count_folds returns total number of folds in a cvo object.

Function cvo_get_inds extracts indices of indicated set (either training or test) from cross-validation object (cvo) created with function cvo_create_folds when fold of interest is indicated as fold (it can be either a name of fold or an index of fold). If cvo does not contain indices of chosen type, it returns a complement to those indices and this information is indicated in attributes of returned object.

Function cvo_get_seeds returns information about seeds of (pseudo)random number generator used for each repetition of splitting to folds.

See also

cvo_create_folds

Examples

cvo <- cvo_create_folds(fluorescence, folds = 10)
cvo

#> $Rep1_Fold01
#>   [1]   2   3   4   5   6   8   9  10  11  12  13  14  15  17  18  19  20  21
#>  [19]  22  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39  40
#>  [37]  41  42  43  44  45  46  47  48  49  50  51  52  53  54  56  57  59  60
#>  [55]  62  63  64  66  67  68  69  70  71  72  73  74  75  76  77  78  79  80
#>  [73]  81  82  83  84  87  88  90  91  92  93  94  95  96  97  98  99 100 101
#>  [91] 102 103 104 105 106 107 108 112 113 114 115 116 117 118 119 120 122 123
#> [109] 124 125 126 127 128 129 130 131 132 133 134 135 136 137 138 139 140 141
#> [127] 142 143 144 145 146 147 148 149 150
#> 
#> $Rep1_Fold02
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37
#>  [37]  38  39  40  41  42  43  44  45  47  48  49  50  51  52  53  54  55  57
#>  [55]  58  59  60  61  62  63  64  65  66  67  68  69  71  72  73  74  75  76
#>  [73]  77  78  79  81  82  83  85  86  87  88  89  90  91  92  93  94  95  96
#>  [91]  97  98  99 100 101 102 103 104 105 106 107 108 109 110 111 112 114 115
#> [109] 116 117 118 119 120 121 122 123 124 125 126 127 128 129 131 132 133 136
#> [127] 137 140 141 142 143 145 146 148 149
#> 
#> $Rep1_Fold03
#>   [1]   1   2   3   4   5   7   8   9  10  11  12  13  14  15  16  17  18  19
#>  [19]  20  21  22  23  24  25  26  27  30  31  32  33  34  36  37  38  39  40
#>  [37]  41  42  43  44  45  46  47  49  50  51  52  53  54  55  56  57  58  59
#>  [55]  60  61  62  63  64  65  67  68  69  70  71  72  74  75  76  77  78  79
#>  [73]  80  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  99
#>  [91] 100 101 102 103 105 106 108 109 110 111 112 113 114 115 116 117 118 119
#> [109] 120 121 122 123 124 126 127 130 131 132 133 134 135 136 137 138 139 140
#> [127] 141 142 143 144 146 147 148 149 150
#> 
#> $Rep1_Fold04
#>   [1]   1   3   4   5   6   7   8   9  10  11  13  14  15  16  17  18  19  20
#>  [19]  21  22  23  24  26  27  28  29  30  31  32  33  34  35  36  37  38  39
#>  [37]  40  42  43  44  45  46  47  48  50  51  52  53  54  55  56  57  58  59
#>  [55]  60  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76  77
#>  [73]  78  79  80  82  84  85  86  87  88  89  90  92  93  94  96  97  98 100
#>  [91] 102 103 104 106 107 108 109 110 111 112 113 114 115 116 117 118 119 120
#> [109] 121 122 123 124 125 126 128 129 130 131 132 133 134 135 136 138 139 140
#> [127] 141 143 144 145 146 147 148 149 150
#> 
#> $Rep1_Fold05
#>   [1]   1   2   4   5   6   7   8  10  11  12  13  15  16  17  18  19  20  21
#>  [19]  22  23  24  25  26  27  28  29  30  31  32  33  35  37  38  39  40  41
#>  [37]  42  43  44  45  46  47  48  49  50  52  53  54  55  56  57  58  59  60
#>  [55]  61  62  63  64  65  66  67  68  70  71  72  73  74  75  76  78  79  80
#>  [73]  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98
#>  [91]  99 101 102 104 105 107 108 109 110 111 112 113 114 116 117 119 120 121
#> [109] 122 123 124 125 126 127 128 129 130 131 133 134 135 136 137 138 139 140
#> [127] 141 142 143 144 145 146 147 148 150
#> 
#> $Rep1_Fold06
#>   [1]   1   2   3   4   6   7   8   9  11  12  13  14  15  16  17  18  19  22
#>  [19]  23  25  26  27  28  29  30  31  32  33  34  35  36  37  39  40  41  42
#>  [37]  43  44  45  46  47  48  49  51  52  53  54  55  56  57  58  59  60  61
#>  [55]  62  65  66  67  69  70  71  72  73  75  76  77  78  79  80  81  82  83
#>  [73]  84  85  86  87  88  89  90  91  93  94  95  96  97  98  99 100 101 103
#>  [91] 104 105 106 107 108 109 110 111 112 113 114 115 116 117 118 119 120 121
#> [109] 122 123 124 125 126 127 128 129 130 131 132 134 135 137 138 139 140 141
#> [127] 142 143 144 145 146 147 148 149 150
#> 
#> $Rep1_Fold07
#>   [1]   1   2   3   4   5   6   7   8   9  10  12  14  15  16  17  18  19  20
#>  [19]  21  22  23  24  25  26  27  28  29  30  31  32  34  35  36  38  39  40
#>  [37]  41  44  46  47  48  49  50  51  52  54  55  56  58  59  61  62  63  64
#>  [55]  65  66  67  68  69  70  71  72  73  74  76  77  78  79  80  81  82  83
#>  [73]  84  85  86  87  89  90  91  92  93  94  95  96  97  98  99 100 101 102
#>  [91] 103 104 105 106 107 108 109 110 111 112 113 114 115 116 117 118 120 121
#> [109] 123 124 125 127 128 129 130 131 132 133 134 135 136 137 138 139 140 141
#> [127] 142 143 144 145 146 147 148 149 150
#> 
#> $Rep1_Fold08
#>   [1]   1   2   3   5   6   7   8   9  10  11  12  13  14  16  17  19  20  21
#>  [19]  22  23  24  25  26  28  29  30  31  32  33  34  35  36  37  38  39  40
#>  [37]  41  42  43  45  46  47  48  49  50  51  53  55  56  57  58  59  60  61
#>  [55]  62  63  64  65  66  68  69  70  71  72  73  74  75  76  77  78  79  80
#>  [73]  81  82  83  84  85  86  87  88  89  91  92  93  94  95  96  97  98  99
#>  [91] 100 101 102 103 104 105 106 107 108 109 110 111 112 113 114 115 116 117
#> [109] 118 119 121 122 124 125 126 127 128 129 130 131 132 133 134 135 136 137
#> [127] 138 139 141 142 144 145 147 149 150
#> 
#> $Rep1_Fold09
#>   [1]   1   2   3   4   5   6   7   9  10  11  12  13  14  15  16  17  18  19
#>  [19]  20  21  22  23  24  25  26  27  28  29  33  34  35  36  37  38  41  42
#>  [37]  43  44  45  46  48  49  50  51  52  53  54  55  56  57  58  59  60  61
#>  [55]  63  64  65  66  67  68  69  70  71  73  74  75  76  77  80  81  82  83
#>  [73]  84  85  86  87  88  89  90  91  92  93  95  97  98  99 100 101 102 103
#>  [91] 104 105 106 107 108 109 110 111 112 113 114 115 117 118 119 120 121 122
#> [109] 123 124 125 126 127 128 129 130 131 132 133 134 135 136 137 138 139 140
#> [127] 142 143 144 145 146 147 148 149 150
#> 
#> $Rep1_Fold10
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  18  20
#>  [19]  21  22  23  24  25  27  28  29  30  31  32  33  34  35  36  37  38  39
#>  [37]  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57
#>  [55]  58  60  61  62  63  64  65  66  67  68  69  70  72  73  74  75  77  78
#>  [73]  79  80  81  83  84  85  86  88  89  90  91  92  94  95  96  97  98  99
#>  [91] 100 101 102 103 104 105 106 107 109 110 111 113 115 116 118 119 120 121
#> [109] 122 123 125 126 127 128 129 130 132 133 134 135 136 137 138 139 140 141
#> [127] 142 143 144 145 146 147 148 149 150
#> 
#> attr(,"class")
#> [1] "cvo_caret" "cvo"      
#> attr(,"info")
#>   indices stratified blocked cv_type  k repetitions sample_size
#> 1   Train      FALSE   FALSE  k-fold 10           1         150
#> attr(,"seeds")
#> attr(,"seeds")$generator
#> [1] "L'Ecuyer-CMRG"
#> 
#> attr(,"seeds")$seeds
#> [1] 3687206
#> 

cvo_get_info(cvo)

#>   indices stratified blocked cv_type  k repetitions sample_size
#> 1   Train      FALSE   FALSE  k-fold 10           1         150

cvo_get_seeds(cvo)

#> $generator
#> [1] "L'Ecuyer-CMRG"
#> 
#> $seeds
#> [1] 3687206
#> 

cvo_get_sample_size(cvo)

#> [1] 150

cvo_count_folds(cvo)

#> [1] 10

cvo_get_fold_names(cvo)

#>  [1] "Rep1_Fold01" "Rep1_Fold02" "Rep1_Fold03" "Rep1_Fold04" "Rep1_Fold05"
#>  [6] "Rep1_Fold06" "Rep1_Fold07" "Rep1_Fold08" "Rep1_Fold09" "Rep1_Fold10"

cvo_get_inds(cvo, 1)

#>   [1]   2   3   4   5   6   8   9  10  11  12  13  14  15  17  18  19  20  21
#>  [19]  22  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39  40
#>  [37]  41  42  43  44  45  46  47  48  49  50  51  52  53  54  56  57  59  60
#>  [55]  62  63  64  66  67  68  69  70  71  72  73  74  75  76  77  78  79  80
#>  [73]  81  82  83  84  87  88  90  91  92  93  94  95  96  97  98  99 100 101
#>  [91] 102 103 104 105 106 107 108 112 113 114 115 116 117 118 119 120 122 123
#> [109] 124 125 126 127 128 129 130 131 132 133 134 135 136 137 138 139 140 141
#> [127] 142 143 144 145 146 147 148 149 150
#> attr(,"fold")
#> [1] "Rep1_Fold01"
#> attr(,"indices")
#> [1] "Train"

cvo_get_inds(cvo, 1, "train")

#>   [1]   2   3   4   5   6   8   9  10  11  12  13  14  15  17  18  19  20  21
#>  [19]  22  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39  40
#>  [37]  41  42  43  44  45  46  47  48  49  50  51  52  53  54  56  57  59  60
#>  [55]  62  63  64  66  67  68  69  70  71  72  73  74  75  76  77  78  79  80
#>  [73]  81  82  83  84  87  88  90  91  92  93  94  95  96  97  98  99 100 101
#>  [91] 102 103 104 105 106 107 108 112 113 114 115 116 117 118 119 120 122 123
#> [109] 124 125 126 127 128 129 130 131 132 133 134 135 136 137 138 139 140 141
#> [127] 142 143 144 145 146 147 148 149 150
#> attr(,"fold")
#> [1] "Rep1_Fold01"
#> attr(,"indices")
#> [1] "train"

cvo_get_inds(cvo, 1, "test")

#>  [1]   1   7  16  23  55  58  61  65  85  86  89 109 110 111 121
#> attr(,"fold")
#> [1] "Rep1_Fold01"
#> attr(,"indices")
#> [1] "Complement to Train"