#' [!] Carry out the multiROC analysis with cross-validation

#' [!] Carry out the multiROC analysis with cross-validation

roc_multiroc_cv(x, gr, optimize_by = "bac", cvo = cvo_create_folds(x, gr,
  seeds = seeds, kind = kind), seeds = NULL, kind = "L'Ecuyer-CMRG")

Arguments

x
A numeric matrix, a data frame, a hyperSpec object or another type of object, convertible to a numeric matrix.

gr
Either a string (scalar, character(1)) or a factor variable (a vector) which defines groups in x.

optimize_by
(string(1)) A string with the name of classification performance measure to use. Currently available options:

"bac" - for balanced accuracy (mean of sensitivity and specificity);

"kappa" - for Cohens kappa;

"youden" - for Youden's index;

cvo
a cross-validation object (cvo), created with function cvo_create_folds, caret createFolds or similar.

seeds
(vector of integers | NULL) Seeds for random number generator for each repetition. If seeds = NULL random seeds are generated. If number of repetitions is greater than number of provided seeds, random seeds are generated and added to the provided ones. The first seed will be used to ensure reproducibility of the randomly generated seeds.

(See set.seed for more information about random number generation).

kind
(character | NULL) The kind of (pseudo)random number generator. Default is "L'Ecuyer-CMRG" as it provides the basis for the multiple streams used in package parallel. More information at set.seed.

Value

A data frame with results. The object also inferits from class multiroc_cv (for displaying purposes).

Examples

library(multiROC)

# --- For numeric vectors objects ---

data(PlantGrowth)

roc_multiroc_cv(PlantGrowth$weight, PlantGrowth$group)

#>          set       fold compared_groups feature median_neg cutoff median_pos
#> 1   training Rep1_Fold1   ctrl vs. trt1       1       4.88   4.23       4.75
#> 2   training Rep1_Fold1   ctrl vs. trt2       1       4.88   5.22       5.33
#> 3   training Rep1_Fold1   trt1 vs. trt2       1       4.75    4.9       5.33
#> 4   training Rep1_Fold2   ctrl vs. trt1       1       5.15   4.37       4.55
#> ...      ...        ...             ...     ...        ...    ...        ...
#> 27      test Rep1_Fold4   trt1 vs. trt2       1         NA    4.9         NA
#> 28      test Rep1_Fold5   ctrl vs. trt1       1         NA   4.35         NA
#> 29      test Rep1_Fold5   ctrl vs. trt2       1         NA   5.22         NA
#> 30      test Rep1_Fold5   trt1 vs. trt2       1         NA   4.87         NA
#>                                                                             
#>      tp  fn  fp  tn sens spec  ppv  npv  bac youden kappa  auc
#> 1     3   5   1   7 0.38 0.88 0.75 0.58 0.62   0.25  0.25 0.57
#> 2     6   2   2   6 0.75 0.75 0.75 0.75 0.75   0.50  0.50 0.77
#> 3     8   0   2   6 1.00 0.75 0.80 1.00 0.88   0.75  0.75 0.78
#> 4     4   4   0   8 0.50 1.00 1.00 0.67 0.75   0.50  0.50 0.69
#> ... ... ... ... ...  ...  ...  ...  ...  ...    ...   ...  ...
#> 27    2   0   0   2 1.00 1.00 1.00 1.00 1.00   1.00  1.00   NA
#> 28    0   2   0   2 0.00 1.00  NaN 0.50 0.50   0.00  0.00   NA
#> 29    2   0   1   1 1.00 0.50 0.67 1.00 0.75   0.50  0.50   NA
#> 30    2   0   2   0 1.00 0.00 0.50  NaN 0.50   0.00  0.00   NA
#>                                                               

roc_multiroc_cv(PlantGrowth$weight, gl(2, 1, 30))

#>         set       fold compared_groups feature median_neg cutoff median_pos tp
#> 1  training Rep1_Fold1         1 vs. 2       1       5.25   5.27       5.13  9
#> 2  training Rep1_Fold2         1 vs. 2       1       5.25   4.09        5.2  2
#> 3  training Rep1_Fold3         1 vs. 2       1       5.17   5.25          5 10
#> 4  training Rep1_Fold4         1 vs. 2       1       5.25   4.88       4.79  7
#> 5  training Rep1_Fold5         1 vs. 2       1       5.17   3.66       5.13  2
#> 6      test Rep1_Fold1         1 vs. 2       1         NA   5.27         NA  2
#> 7      test Rep1_Fold2         1 vs. 2       1         NA   4.09         NA  1
#> 8      test Rep1_Fold3         1 vs. 2       1         NA   5.25         NA  1
#> 9      test Rep1_Fold4         1 vs. 2       1         NA   4.88         NA  0
#> 10     test Rep1_Fold5         1 vs. 2       1         NA   3.66         NA  0
#>    fn fp tn sens spec  ppv  npv  bac youden kappa  auc
#> 1   3  6  6 0.75 0.50 0.60 0.67 0.62   0.25  0.25 0.55
#> 2  10  0 12 0.17 1.00 1.00 0.55 0.58   0.17  0.17 0.55
#> 3   2  7  5 0.83 0.42 0.59 0.71 0.62   0.25  0.25 0.62
#> 4   5  4  8 0.58 0.67 0.64 0.62 0.62   0.25  0.25 0.61
#> 5  10  0 12 0.17 1.00 1.00 0.55 0.58   0.17  0.17 0.55
#> 6   1  2  1 0.67 0.33 0.50 0.50 0.50   0.00  0.00   NA
#> 7   2  0  3 0.33 1.00 1.00 0.60 0.67   0.33  0.33   NA
#> 8   2  1  2 0.33 0.67 0.50 0.50 0.50   0.00  0.00   NA
#> 9   3  1  2 0.00 0.67 0.00 0.40 0.33  -0.33 -0.33   NA
#> 10  3  0  3 0.00 1.00  NaN 0.50 0.50   0.00  0.00   NA